Create refactor.rb
